requirements.txt
appdirs==1.4.3
attrs==19.1.0
blis==0.2.4
cachetools==3.1.0
certifi==2019.3.9
chardet==3.0.4
cymem==2.0.2
en-vectors-web-lg==2.1.0
freeze==1.0.10
google-api-core==1.9.0
google-auth==1.6.3
google-cloud==0.34.0
google-cloud-core==0.29.1
google-cloud-storage==1.15.0
google-resumable-media==0.3.2
googleapis-common-protos==1.5.9
idna==2.8
importlib-metadata==0.9
jsonschema==2.6.0
lexrank==0.1.0
murmurhash==1.0.2
nltk==3.4.1
numpy==1.16.2
pandas==0.24.2
path.py==12.0.1
pkg-resources==0.0.0
plac==0.9.6
preshed==2.0.1
protobuf==3.7.1
pyasn1==0.4.5
pyasn1-modules==0.2.4
pyrsistent==0.14.11
python-dateutil==2.8.0
pytz==2019.1
pywsd==1.1.7
regex==2019.4.14
requests==2.21.0
rsa==4.0
scipy==1.2.1
six==1.12.0
spacy==2.1.3
srsly==0.0.5
summa==1.2.0
thinc==7.0.4
tqdm==4.31.1
uritools==2.2.0
urlextract==0.10
urllib3==1.24.2
wasabi==0.2.1
zipp==0.3.3
